+++ src/SS2KLog.cpp
@@ -115,6 +115,14 @@ int ss2k_log_hex_to_buffer(const byte *data, const size_t data_length, char *buf
   return written;
 }
 
+int ss2k_log_hex_to_buffer(const char *data, const size_t data_length, char *buffer, const int buffer_offset, const size_t buffer_length) {
+  int written = 0;
+  for (int data_offset = 0; data_offset < data_length; data_offset++) {
+    written += snprintf(buffer + buffer_offset + written, buffer_length - written + buffer_offset, "%02x ", *(data + data_offset));
+  }
+  return written;
+}
+
 void ss2k_log_write(esp_log_level_t level, const char *module, const char *format, ...) {
   va_list args;
   va_start(args, format);